Added Tests
Added DXFP sample that uses both clock time with fraction seconds and
clock time with frames.

+++ pycaption/dfxp.py
@@ -103,7 +103,7 @@ class DFXPReader(BaseReader):
         microseconds = int(int(timesplit[0]) * 3600000000 + int(timesplit[1]) * 60000000 + (float(timesplit[2])*1000000))
         # are there frames? - add them assuming video is 30 frames/second
         if len(timesplit) > 3:
-            microseconds = microseconds + ((float(timesplit[3]) / 29.97) * 1000000)
+            microseconds = microseconds + int((float(timesplit[3]) / 29.97) * 1000000)
         return microseconds
 
     def _translate_tag(self, tag):
